requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
cmake==3.26.1
filelock==3.11.0
idna==3.4
Jinja2==3.1.2
lit==16.0.0
MarkupSafe==2.1.2
mpmath==1.3.0
networkx==3.1
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
pandas==2.0.0
Pillow==9.5.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.28.2
six==1.16.0
sklearn==0.0.post1
sympy==1.11.1
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
triton==2.0.0
typing-extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
